splitting balance services
