+++ src/services/balances/L2LPPending/index.ts
@@ -1,0 +1,6 @@
+import { BigNumber } from 'ethers'
+
+export const L2LPPending = () => {
+  const sum = BigNumber.from('0')
+  return sum.toString()
+}
